feat: remove "character" field that was never impl.
I believe the intention was that this field should specify the characters that were somehow involved, but it was never implemented.

I think a much better solution is to simply infer what characters are involved in a mission by looking at the character sheets exhaustively, which is of course easy to do programatically. Those sheets specify every mission the character is part of anyway so that won't be a problem. Besides, two-way relationships are just difficult to maintain.

When these sanity checks are removed from the "mission" and "event" resources there's currently nothing else to check, so I'm removing that boilerplate.